Move implementation into the controller
