+++ apps/files_sharing/lib/AppInfo/Application.php
@@ -28,7 +28,6 @@
 namespace OCA\Files_Sharing\AppInfo;
 
 use OC\AppFramework\Utility\SimpleContainer;
-use OCA\Files_Sharing\API\Share20OCS;
 use OCA\Files_Sharing\Controller\Share20OcsController;
 use OCA\Files_Sharing\Controllers\ExternalSharesController;
 use OCA\Files_Sharing\Controllers\ShareController;
@@ -78,12 +77,13 @@ class Application extends App {
 			);
 		});
 
-		$container->registerService('Share20OcsApi', function (SimpleContainer $c) use ($server) {
-			return new Share20OCS(
+		$container->registerService('Share20OcsController', function (SimpleContainer $c) use ($server) {
+			return new Share20OcsController(
+				$c->query('AppName'),
+				$server->getRequest(),
 				$server->getShareManager(),
 				$server->getGroupManager(),
 				$server->getUserManager(),
-				$server->getRequest(),
 				$server->getRootFolder(),
 				$server->getURLGenerator(),
 				$server->getUserSession()->getUser(),
@@ -92,14 +92,6 @@ class Application extends App {
 				$c->query(NotificationPublisher::class),
 				$server->getEventDispatcher(),
 				$c->query(SharingBlacklist::class)
-			);
-		});
-
-		$container->registerService('Share20OcsController', function (SimpleContainer $c) use ($server) {
-			return new Share20OcsController(
-				$c->query('AppName'),
-				$server->getRequest(),
-				$c->query('Share20OcsApi')
 			);
 		});
